Move javascript test-config into tests dir
