+++ javascript/tests/small/error-handling.test.js
@@ -3,7 +3,7 @@
  */
 
 import WaveClient, { WaveClientError, RateLimitError } from '../../src/wave-client.js';
-import { MOCK_DATA, createMockFetch } from '../../test-config.js';
+import { MOCK_DATA, createMockFetch } from '../test-config.js';
 import { TestSetup } from '../test-utils.js';
 
 describe('WaveClient Error Handling', () => {
