+++ javascript/tests/small/experiment-data.test.js
@@ -10,7 +10,7 @@ import WaveClient, {
     NotFoundError,
     ServerError
 } from '../../src/wave-client.js';
-import { MOCK_DATA, createMockFetch } from '../../test-config.js';
+import { MOCK_DATA, createMockFetch } from '../test-config.js';
 import { TestSetup } from '../test-utils.js';
 
 describe('WaveClient logExperimentData', () => {
